added logic in counter aspect

+++ src/main/java/com/epam/training/spring/aspects/ShowingCounter.java
@@ -4,6 +4,6 @@ package com.epam.training.spring.aspects;
 import com.epam.training.spring.domains.Showing;
 
 public interface ShowingCounter {
-    int getAccessCount(Showing showing);
-    int getBookCount(Showing showing);
+    long getAccessCount(Showing showing);
+    long getBookCount(Showing showing);
 }
